Test/update ci test (#109)
* test: add varlen backward tests for chunk_simple_gla

Add test_simple_gla_varlen_bwd (12 cases: 3 gates x 4 shapes) and
test_simple_gla_varlen_bwd_with_h0 to cover the previously untested
varlen backward path. Reference is per-sequence jax.grad(naive) at fp64.

Config T100 (segs=[30,70], C=64) and T65 (segs=[1,64], C=64) exercise
partial chunks, validating the dg_last placement fix.


* update simple ci test.

* add ci test

* update

* [WIP] Fix Mosaic TPU kernel compilation errors in chunk_o.py (#112)

* Initial plan

* fix: explicit float32 cast for dot inputs in chunk_o.py Pallas kernels

Agent-Logs-Url: https://github.com/primatrix/pallas-kernel/sessions/516490a2-aab3-4006-9bd5-a0ed78850fb8


---------


* Update chunk_o.py with new content

* fix: use exp(g_i - g_j) difference pattern in simple_gla backward kernel for numerical stability (#116)

The backward kernel in chunk_simple_gla_bwd_kernel used exp(-g) and exp(g)
separately, which overflows in float32 for large negative g values (e.g.,
g_gamma=-5, position=18 → exp(90) = inf). Replace with the stable exp(g_i - g_j)
difference pattern where the exponent is always ≤ 0. Also fix g_gamma generation
in GPU tests (torch.sigmoid → -F.softplus) to ensure negative log-decay values.


* update simple ci test.

* Update chunk_o.py with new content

* fix: use exp(g_i - g_j) difference pattern in simple_gla backward kernel for numerical stability (#116)

The backward kernel in chunk_simple_gla_bwd_kernel used exp(-g) and exp(g)
separately, which overflows in float32 for large negative g values (e.g.,
g_gamma=-5, position=18 → exp(90) = inf). Replace with the stable exp(g_i - g_j)
difference pattern where the exponent is always ≤ 0. Also fix g_gamma generation
in GPU tests (torch.sigmoid → -F.softplus) to ensure negative log-decay values.


* test: add varlen backward tests for chunk_simple_gla

Add test_simple_gla_varlen_bwd (12 cases: 3 gates x 4 shapes) and
test_simple_gla_varlen_bwd_with_h0 to cover the previously untested
varlen backward path. Reference is per-sequence jax.grad(naive) at fp64.

Config T100 (segs=[30,70], C=64) and T65 (segs=[1,64], C=64) exercise
partial chunks, validating the dg_last placement fix.


* update

* fix

* fix ci error

* fix ci error

* fix: widen dg tolerance for varlen bwd GPU test

At chunk/segment boundaries the dg gradient is near-zero, causing
Triton vs JAX CPU rounding divergence in dg_last accumulation.
Raise dg atol to 1e-3 (from 5e-4) to accommodate this.


---------

+++ tops/cpu/ops/common/chunk_o.py
@@ -203,7 +203,8 @@ def chunk_fwd_o(
         b_A = b_A * jnp.exp(b_g_gamma[None, :, :, None] - b_g_gamma_key[None, None, :, :])
 
       b_A = jnp.where(causal_mask[None, :, None, :], b_A, 0)
-      b_o = b_o * scale + dot("bihj,bjhv->bihv", b_A.astype(v.dtype), b_v, acc) * scale
+      # Keep b_A in fp32 for precision; upcast b_v instead.
+      b_o = b_o * scale + dot("bihj,bjhv->bihv", b_A, b_v.astype(jnp.float32), acc) * scale
 
       o_buf = write_chunk(o_buf, b_o.astype(v.dtype), start, valid_len)
 
@@ -248,7 +249,8 @@ def chunk_fwd_o(
       b_A = b_A * jnp.exp(b_g_gamma[None, :, :, None] - b_g_gamma_key[None, None, :, :])
 
     b_A = jnp.where(causal_mask[None, :, None, :], b_A, 0)
-    b_o = b_o * scale + dot("bihj,bjhv->bihv", b_A.astype(v.dtype), b_v, acc) * scale
+    # Keep b_A in fp32 for precision; upcast b_v instead.
+    b_o = b_o * scale + dot("bihj,bjhv->bihv", b_A, b_v.astype(jnp.float32), acc) * scale
 
     o_chunks.append(b_o.astype(v.dtype))
